fixed 100% CPU spinning on HP-UX upon a connection break
As I see the problem seems to be caused by the fact that HP-UX returns
POLLERR only without the other bits (e.g. POLLHUP) syslog-ng would
handle this gracefully if either the other bits would be set, or there'd
be some pending messages to send, in which case a normal write() error
would occur.

Conflicts:

	src/logwriter.c

Conflicts:

	src/logwriter.c