[LogReader] only assume that a file was moved if the size of the file is non-zero
If an external logrotate program is used to rotate a logfile, a
small race still exists when using syslog-ng to read that logfile,
as described by Evan Rempel:

"
1. Application is writing to log file named "A".
2. External log rotation renames "A" to "A.1"
3. External log rotation touches/creates file named "A" and sets appropriate
    permissions.
4. Internal timer of syslog-ng is triggered by follow_freq() setting. Syslog-ng will
    switch to the new file "A" because it exists, even though it was created only
    milliseconds earlier.
5. log rotation signals the application to switch log files (reload or restart).
6. Application flushes log buffers to current file which is now A.1, but syslog-ng
    is no longer reading this file.
7. Application closes current log file "A.1" and opens new log file "A".

This sequence will result in the last buffer flush (step 6) from the application to
be missed by syslog-ng.
"

This patch makes syslog-ng to switch to the new log file if it already
received some data.