fixed an uninitialized variable warning
The warning message is right, when a non-IPv4 & non-UNIX peer sends a
message is received, then the "netmask()" filter uses an uninitialied
variable. However the function probably was returning the right value even
in this case as well by accident.

filter.c:452: warning: 'addr.s_addr' may be used uninitialized in this function

Reported-By: Chanchal Verma <cverma2@csc.com>