[main] never initialize the configuration in the supervisor process
Earlier versions of syslog-ng 3.0 initialized the configuration in the
supervisor process (as root) on non-Linux platforms. However this
causes trouble on FreeBSD: /dev/klog is kept opened in the supervisor
which prevents the child process to open the device.

Also, the fds initialized early during startup might be completely bogus
by the time syslog-ng crashes and the supervisor relaunches syslog-ng.
(e.g. the fd to /dev/log might already be stale). Thus this behaviour to
support initializing the configuration as root is changed: now everything
is done as a non-privileged user if that is requested by the user
(with -u/-g command line options).

This is somewhat incompatible but I can see no other solutions (except maybe
supporting other capability-like systems on other platforms).