kill the syslog-ng process if reverting to the old config fails
When the configuration is reloaded, syslog-ng attempts to init the new
configuration and reverts to the old one in case it fails.

However reverting to the old one can fail too, this case was not
really handled previously. Now syslog-ng tries to kill itself
using SIGQUIT, in order to have itself relaunched by the supervisor.

There's really nothing else to be done here.