fix file descriptor leak when using includes
When using the newly introduced includes feature the files that are
included are not closed and cause a file descriptor leak.
Depending on the system ulimit, number of includes and frequency of
re-configuration this can cause configuration not to be applied.

To address this it is necessary to close the included file after
processing it.

Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>