[config parser] don't use facility/level keywords for specifying the default values for file sources
There was an experimental feature to add support for controlling the
default facility/priority values for sources in case the incoming
message lacked this value.

This used the same 'facility' and 'level' keywords but this was misleading
as it does not override the value received if it was present.

This patch changes it to use distinct keywords.