timezone related variable naming cleaned up for readability, added support for local_timezone() (fixes: pub#44)
This patch clarifies time_zone related variables by adding "recv" or "send"
prefix to indicate the information that was previously implied.

Also added a 3rd time zone named "local" which controls the timezone to
be used for naming files/SQL tables. The local timezone is supported by
the SQL and the file destination.