INCOMPATIBLE: fixed command line argument incompatibility with syslog-ng 2.x
syslog-ng 3.0 was changed to use the command line parser from Glib instead
of the system getopt() which does not support long arguments on some
platforms.

However there was an unintended conflict in the processing of the -C
(chroot) and -R (persist-file) options, which caused an unintended change of
meanings of these options.

Since these options are quite rarely used and it is expected that more users
are yet to be upgraded to 3.0, it was decided to keep the compatibility
since 2.x, thus it is an incompatible change within the 3.0 series.

Reported-By: Leonid Chaichenets