[timestamp] solving the daylight saving problem (fixes: #20182)
The problem was that the local services (CRON, dnsmasq) sends BSD timestamp
which hasn't any zone information and we assumed that there wasn't daylight
saving, so set the tm.tm_isdst to -1.

The solution is that let's assume that the correct tm_isdst value is what we
get from the current local time.

If timezone isn't set, use the local one
and calculate the correct time