Implement NVTable to store name-value pairs
This patch is the first of the reorganization to use the same data structure
API to access all string-like fields of a LogMessage. Instead of having
4 different implementation for:

  * standard fields of a log message (HOST, PROGRAM, MESSAGE, etc)
  * regexp matches ($1, $2, ...)
  * structured data
  * name-value pairs

This patchset aims to uniformize all of these while still keeping the
performance up. It is believed (and some evidence has shown) that
accessing standard fields is not slower, while name-value pairs become much
faster (up to 4 times speedup). It is just an added bonus that all
client code can use the same APIs, thus the difference between name-value
pairs and macros can fade even more.