pdbtool: fixed a possible format string bug
one of the g_string_append_printf()s appended a "buff" using the
format string argument, which is obviously not very good.

This has no security implications as pdbtool is ran as a user, but could
be a problem if pdbtool is used behind a web application, feeding data
unchecked to pdbtool.