implement SDATA support on top of NVTable
Structured data is a little bit more than simple name-value pairs in the new
LogMessage/NVTable model as it requires proper ordering in order to
reconstruct the structured data field in the log message.

This is accomplished by using an array in the LogMessage that contains the
handles of the SD elements in the original order. Thus the original
SDATA field can be recovered even after putting them in NVTable.

This algorithm is about 6% slower in microbenchmarks than the original
with linked lists, but overall it should be a win because of the less
number of memory allocations and the more cache-friendly layout. I still
need to run a test or two to confirm that though.