NVRegistry: change memory layout to avoid one pointer deref
Since the NVHandle flags are going to be needed in the log_msg_get_value()
fast path to support macros as values, I want to avoid the extra deref. Also
functions taking part in the fast path are moved to the header and
marked as inline.

Among other things this means that the handles are stored in a GArray instead
of a GPtrArray.

This patch also contains some function renames for consistency and the
extension of the handle flags to 16 bits.