LogParser: make sure the LM_V_MESSAGE pointer remains valid accross parsing
Since parsers add values to the LogMessage being parsed, it is perfectly
normal to have their payload reallocated in case the parsed values do not
fit into the currently allocated payload.

log_parser_process() passed a borrowed reference to the parse function,
which in turn could be invalidated in case this reallocation took place.

The solution is to add a reference to msg->payload, and unref it at the end
of the parse call, which makes it unnecessary to copy the message.