fixed rewrite rule processing
In some cases rewritten messages could propagate over branches where they
shouldn't have been changed.

This patch introduces a new message flag "LF_REFERENCED" that indicates that
a given log message was cloned _or_ is sitting on a message queue waiting
to be processed. Whenever we encounter such a message, we always clone that
instead of refing it.