fixed SDATA handling in case the SD-ID or SD-PARAM uses a dot in its name
Enterprise assigned SDATA fields can have dots in their names, which prevent
the proper formatting of structured data fields.

I've used a trick in this case: since the ${.SDATA.xxx.yyy} format used
to reference structured data fields becomes ambigous once xxx or yyy contain
a dot, we need to store the length of the SD-ID. This is now stored
in the NVHandle flags space.

The problematic case is for example:

.SDATA.Events.1234.Param

In this case it is not possible to tell from the value name alone whether
it is "Events.1234" . "Param", or "Events" . "1234.Param", however it is
not realistic to have a real-life collision.