[sdata parser] fixed off-by-one overflows in structured data parser
When parsing SDATA ID, param and value from the input an off-by-one
occurs which means that an extra zero byte is written on the stack.

This is clearly a possible DoS situation, but further exploitation is
not believed to be possible.

The overwritten areas on x86 CPUs:
  * param becomes a zero string if value is overflown
  * ID becomes a zero string if param is overflown
  * the LSB byte of a read-only pointer becomes zero if ID is overflown

Other CPUs may use a different stack ordering.

Reported-By: András Tevesz