grammer/lexer: Generalize parser/lexer code, generalize them to make it more modular
This patch introduces CfgParser and CfgLexer classes which lay
the foundations for a more modular configuration file parser that makes
it possible for plugins to carry their own configuration file grammar.

This patch was broken down from a larger patch set, thus may or may not
compile, the primary reason for committing it separately is to make it
easier to review.