cfg-lex.l: fixed hexadecimal number processing and added \xFF format character literals
There was support for hexadecimal/octal numbers in the lexers, but in
reality the hexadecimal one didn't work for digits in the 'A' - 'F' range.

Also, since there was already requests I've added support for explicitly
specifying characters by their numeric value. The syntax supported is similar
what we have in C/Python:

  \xhh to encode a hexadecimal value
  \ooo to encode an octal value