blocks: make it possible to use any number of ';' between drivers
When including a block it is quite easy to miss whether we need to
include/omit a semicolon in a given context. This problem is somewhat
mitigated by allowing several ';' characters between drivers.

This patch solves this problem for source/destination/rewrite/parser/filter

This also extends the filter syntax to make it possible to use ';' as
an AND operator with low precedence. E.g. expressions between ';'
are evaluated first and at the end they are ANDed together, just
like with filter references embedded in log statements.