cfg-parser: move the keyword table to the lexer context
This patch is a refactorization that adds the keyword table to the
CfgLexerContext, thus a push_context() call can add further keywords
to be recognized. The keyword lookup code now considers all stacked
keywords, instead of only two of them (the context specific
and the global).

This patch also removes _set_current_keywords() and moves the automatic
addition of new keywords to the cfg_parser_parse() call, instead
of the parser specific lexer function.

This patch also contains a minor plugin.c change (to make it possible to
look up Plugin first, then instantiate it in a separate function call)
for improved error reporting.