don't cache the local hostname accross SIGHUPs
Reported by the SUSE maintainer, syslog-ng caches
the local hostname accross SIGHUPs, which causes problems when
dynamic hostname configuration is in use (e.g. DHCP). This patch
relookups the local hostname when a SIGHUP is received.

Also, there was a bug in the per-source use_fqdn() processing which
is fixed by this patch: earlier, if two sources used use_fqdn()
differently the one which received the first message won and the
hostname was set according to its settings, even if the message
was received on the other source.