logmsg: always store the legacy message header (MSGHDR)
This patch essentially makes the 'store-legacy-msghdr' behaviour
the default and creates a new 'dont-store-legacy-msghdr' flag to return
to the 3.0 compatible behaviour.

This patch introduces some incompatibility, however I feel that since the
'store-legacy-msghdr' flag is the assumed behaviour, it is best to make it
default. The only reason it existed in 3.0 is the expected performance
drop, but the 3.1 NVTable refactorizations solved that issue.

It is also needed for the follow-up patch to work properly.