logmsg: don't expect a hostname for local transports in the log message
syslog-ng expected the same message format for both local and network
transports. However this seems to cause problems for applications that
embed a space in their $PROGRAM field, e.g.

logger -t "alma korte" message

Since we know that $HOST is always empty for local log transports
(checked for Linux, FreeBSD, AIX, HP-UX and Solaris) we can simply
assume that even if there's a space it is not a hostname, but rather
an incorrectly formatted $PROGRAM.

This will be recognized as:

$PROGRAM = "alma"
$MSGHDR = "alma "
$MSG = "korte: msg"

E.g. if it is formatted with the default template, it becomes the same as it
was received.

Reported-By: Marius Tomaschewski