Make sure that negative values never get resolved using getpwnam/getgrnam
Originally reported by the SUSE maintainer, syslog-ng is sometimes resolving
invalid usernames using getpwnam/getgrnam. Before this patch,
syslog-ng first attempted to resolve usernames using NSS, then if that failed
tried to interpret them as numeric values.

This logic was reversed, now we first check if it is a number, then if that
fails do we proceed with resolving the username using getpwnam.

Once here, I've also did some cleanups to make the bed for a follow-up patch:
  * change all uid_t/gid_t/mode_t variables to "gint" instead of their
    respective types, to make sure that we can represent negative values
    there
  * all occurences which tried to interpret uid/gid/mode information was
    changed to interpret all negative values the same as -1 (do nothing).
    This makes it possible to use further extremal values in the future.

The original report came with a patch, but the original goal
was completely reimplemented using a different means by Balazs Scheidler.

Reported-By: Marius Tomaschewski