introduce a new syntax to skip owner/group/perm setting on a per-file basis
This change was originally an intended behaviour, but due to a design
problem how global options were inherited to a per-destination basis
at the end it wasn't.

The documented behaviour to skip owner/group/perm setting on a given file
is to use "-1" as the value, however this only gets applied if _both_
the per-destination and the global values are "-1".

E.g. if you want something like:

options { owner("adm"); };

destination d_file_def { file("/var/log/default-owner"); };
destination d_file { file("/var/log/no-owner" owner("-1")); };

The intention with a config like this is ambigious:
  * the logical outcome would be to tell syslog-ng not to touch the
    ownership of the "no-owner" file in any way
  * the reality however is that "adm" ownership would be applied to
    "no-owner" as well, since the "-1" value is used internally in syslog-ng
    to indicate that no per-destination value was specified.

The only way to tell syslog-ng not to touch the ownership of a file is to
set the global setting to -1 and either set -1 for the destination, or omit
the setting (which defaults to -1).

Since it would be tricky to fix this without incompatibilities, I've decided
not to make any, and introduced an additional syntax, which I think is
easier to understand than the original trick with "-1":

destination d_file { file("/var/log/no-owner" owner()); };

This means that the file ownership shouldn't be changed regardless of the
global setting. Internally this is equal to "-2", so the above is equivalent to

destination d_file { file("/var/log/no-owner" owner(-2)); };

But that's an implementation detail, you don't need to know.

Please note that this patch depends on the parent patch as well, without
it, it won't function properly.