main: stay in foreground if syntax-check or debug is requested
There's simply no reason to go into the background just to check the
syntax of the configuration file.

Likewise, debug messages may generate feedback loops in the configuration
thus it is not safe to enable them unless internal messages are only
going to the stderr and syslog-ng is configured to be run in
the foreground.

Reported-By: Corey Hickey