logmsg: don't expect a hostname for local transports in the log message
syslog-ng expected the same message format for both local and network
transports and used heuristics to determine whether the message contained
a hostname or not. This seems to cause problems for applications that
embed a space in their $PROGRAM field, e.g.

logger -t "alma korte" message

Since we know that $HOST is always empty for local log transports
(checked for Linux, FreeBSD, AIX, HP-UX and Solaris) we can simply
assume that even if there's a space it is not a hostname, but rather
an incorrectly formatted $PROGRAM.

This will be recognized as:

$PROGRAM = "alma"
$MSGHDR = "alma "
$MSG = "korte: msg"

E.g. if it is formatted with the default template, it becomes the same as it
was received.

This is an incompatible change and syslog-ng warns about it when trying to
use a pre-3.2 configuration.  The old behaviour can be reset by using the
'expect-hostname' and as a bonus it is also possible to customize this
behaviour even for network sources by using the 'no-hostname' flag.

This patch is rather dirty and was implemented by refactoring the
LogReaderOptions/LogParseOptions parsing code.  It'd have been a lot of
work to separate the two, thus they are in the same patch.

Summary of the refactorization:
==============================

Instead of passing the same set of flags through multiple layers, let the
"syslog parse options" be visible from under LogReaderOptions, which gets
rid off a lot of LR_ defines accompanying the same message parsing option
(e.g. LP_LOCAL & LR_LOCAL)

Make the two set of flags completely distinct, LR_ refer to functionality
in the LogReader, LP_ specifies parsing options. LRO_ was completely
removed.

Reported-By: Marius Tomaschewski