report more problems as syntax errors
Previously a syntax error was very sparse and lacked a lot of important
information, therefore a lot of error cases were reported not via
yyerror(), but rather via log messages.

This was changed with the last grammar modularization. This patch converts
various error cases to be checked by a new macro, which triggers a
syntax error if something goes wrong.