LogProto: added queued method
The new queued method indicates that the last message fetched
got successfully queued to the destination FIFO. This is going
to be used as a feedback when recording the current file position.

e.g. if the message gets lost while being delivered to the destination
queue because of a crash, it will be resent again because the current
file position is not recorded in the persistent state until queued is
called.