cfg-lexer: added CFG_KEYWORD_STOP support for keyword lookup
Instead of open-coding the various contexts that do not need keyword
lookup (block contents, block references, etc) introduce a special keyword
defined using the CFG_KEYWORD_STOP macro, that will cause the keyword
lookup mechanism to stop looking for further keyword items.

This means that if a keyword table contains this item, the items in the
parent context are not checked.