config parser: lay the foundation of syslog-ng.conf global variables
syslog-ng.conf wide global variables are introduced by this patch, which
can be referenced from anywhere in the configuration file, using a syntax
similar to the block argument passing mechanism.

This patch only creates the foundations, variable assignment will only
follow in an upcoming patch.

Syntax:
file("`log-directory`/messages.log");

Where `log-directory` is enclosed in backticks and can name:
  1) a block argument if we're in the middle of a block
  2) a global variable (anywhere)
  3) an environment variable (anywhere)

This backtick syntax can be used inside strings, but also in non-strings,
basically the backticks will be expanded before the lexical analysis phase.

In the future the backtick syntax might be extended in the future in the
format: `cmd params`.