pragma-parser: instead of using a hand-crafted parser for pragma items, use a bison generated one
The pragma stuff became more and more complex: it originally served only
to specify the configuration version, then the @module pragma was
introduced and now I want to add support for syslog-ng.conf global
variables with the @define keyword.

This patch reimplements the pervious pragma functionality using a bison
generated parser, which also lifts the restriction that pragma statements
must be the fist lines of the configuration file.

The complete set of pragmas understood are:

@version: x.y
@define global value
@module mod name(value) name2(value2) ...
@include filename

The preferred format of the include statement became @include to indicate
that it's not a usual statement as it can happen anywhere (unlike the
pre-3.2 include which could only be on the topmost level)