SCL: the beginnings of "syslog-ng configuration library" aka SCL
This patch redefines "contributions" for syslog-ng. The work up until
this patch essentially creates a new level of flexibility for syslog-ng.

SCL is a set of scripts, configuration files and other related information
that will make it easier to integrate applications to syslog-ng for
log collection.

Also, it contains a new, default configuration file for syslog-ng, that will
work on all currently supported platforms. This ends an era where you had to
create syslog-ng configuration file from scratch, because it contained
platform specific stuff.

The supplied configuration files will not overwrite the ones you already
created, but in order to use the new framework, it is advisable to check
scl/syslog-ng.conf and include the following in your configuration file:

@version: 3.2
@include 'scl.conf'

With this boilerplate, you get a new "system()" source that will expand
in a platform specific way, as necessary. The actual result of this "macro"
is the output of the shell script /scl/system/generate-system-source.sh
5