added preprocess-into command line argument
This patch adds a new command line argument --preprocess-into which creates
an expanded version of the configuration file (e.g. post-block and define
expansion) in order to make troubleshooting easier.

With the new blocks/plugins framework it may not be so easy to see what
syslog-ng actually processes, this output file makes that easier.