logwriter: fixed flush_timeout handling on non-file destinations
Non-file destinations may not always be willing to accept messages, which
means that even if the flush_timeout elapsed, the destination may not be
ready to accept messages.

However log_writer_fd_prepare() basically assumed that once the timeout
elapses we can immediately start writing messages (which is true for
file destinations but false for others).

This patch fixes this case.

Kudos to Evan who helped me to track down this issue.

Reported-By: Evan Rempel