LogProtoFileReader: new LogProto class that remembers file positions in PersistState
This patch is quite complex and is the most important part of the
atomic file position updates function.

First of all, it separates LogProtoPlainServer state into a separate
structure, in order to make it possible to store in plain memory (sockets)
or in the PersistState file. This state information represents everything
that is needed to reopen the file exactly at the current position.

NOTE that it is not generally possible to ensure that two independent
memory accesses will survive a crash/asynchronous signal, thus it is
possible to kill syslog-ng right in the middle of such a state update
operation. However memory accesses were ordered to ensure that in this
case no message loss occurs, rather syslog-ng will duplicate messages.

This generally means that internal buffer positions are set to 0 _before_
updating the file position. If syslog-ng crashes before the file position
gets updated, then the old buffer will be sent twice (since buffer_pos is
0, while file_pos still refers to the old buffer).

Second, this patch reorganizes state load/save/reset into a simple
restart_with_state method, that allocates the state, restores the file
position and lets the syslog-ng core and the kernel to do the rest
by storing variables in an mmapped memory region.

Third, the persistent state allocation and old persist file conversion
functions are moved to a separate LogProtoFileReader class, which is
basically a LogProtoPlainServer but ensuring that state is in PersistState.

Fourth, this patch changes affile.c to use this infrastructure, thereby
completing the 'atomic-file-position-updates' functionality.