msg-format: make it possible for a format plugin to specify a factory for LogProto
Some format plugins (notably the upcoming pacct) will possibly need a
custom LogProto instance when reading binary file formats. This
patch adds a construct_proto() virtual function to the MsgFormatHandler
class for exactly this purpose.

NOTE: this callback only works for file sources as of now, but may change in
the future.