plugins: make it possible to link a module directly into the executable
With this change, it is possible to link a module directly into the main
executable. This is used for example with unit test programs, but would also
make it possible to create a statically linked syslog-ng binary.

The only change in the modules is that the name of the init function has
changed, from "syslogng_module_init()" to a module specific function:
"<modulename>_module_init()".