separate directory hierarchy into lib/, modules/ and syslog-ng/
It was more and more cumbersome to compile syslog-ng during development, since
a core change always recompiled (or just relinked) the modules as well
even when that was not strictly necessary. (make install in src/)

This patch reorganizes the directory hierarchy, lib is where the core lives
modules where plugins are and syslog-ng is where the main program resides.