db-parser: automatically assign tags to classified messages
Until now the "class" attribute in a patterndb rule only associated a
name-value pair to the message (.classifier.class), with this change
we also associate a tag named '.classifier.<class>' to make it easier
and faster to match against the class of a message in filter expressions.

Also, this patch makes the class attribute optional, in which case
it defaults to "system".