templates: rework the template function system to allow more complex functions & message contexts
This patch changes the template function system to allow:
  1) function specific preprocessing and evaluation rules
  2) support for multiple messages as context

The first one makes it possible to create $(if cond) function where cond
is not a template but rather a syslog-ng filter.

The second makes it possible to reference name-value pairs for messages
other than the current one, used when correllating events.