dbparser: added timerwheel implementation for correllation timers
We need a lot of timers for correllation, basically each correllation
state has its own timer. Calculating with 30k msg/sec means that we have
roughly about 15k/sec timer addition and 15k/sec timer deletion
during.

Therefore it is important to use a timer mechanism that is as fast
as possible. This implementation mimics the timerwheel mechanism of
the Linux kernel.