patterndb: reworked + greatly enhanced how correllation works
An early proof-of-concept correllation implementation was done earlier, this
patch enhances that with a lot of new functionality while changing some of the
details to make room for new stuff.

Correllation between patterndb rules work as follows:

 * <rule> tags now how context-scope, context-id and context-timeout
   attributes
 * once a message matches a rule with a context-id attribute it associated
   with the correllation state identified by context-id (no separate
   join/store like in the earlier implementation).

context-scope is basically a "namespace" that specifies portions of the
session ID that's most probably needed in a lot of cases. These are
the valid values of context-scope:

  * process: correllation is done for messages with the same
    host/program/pid (e.g. the same process). Any messages
    having mismatching values are not even considered.
  * program: correllation between messages of the same host/program
  * host: correllation between messages of the same host
  * global: correlation between the complete message stream,
    no specific filtering

Basically context-scope could be implemented alternatively
using $HOST/$PROGRAM/$PID macros in the context-id, but since these would
often be used, these are built into syslog-ng, and the implementation
is optimized for this. (e.g. faster than having separate macros).

Once a rule matches, additional "actions" can be specified using the <actions>
tag:

<actions>
  <action condition="<filter expr>" rate="<num>/<period>" trigger="<trigger>">
    <message>
      <value name="MESSAGE">foo</value>
      <tags>
        <tag>bar</tag>
      </tags>
    </message>
  </action>
</actions>

Right now the only possible action is to generate a new message, specified
by the <message> tag (in the future other tags can be permitted here).

The attributes of the <action> tag are as follows:
  * condition: specifies a syslog-ng filter expression that needs to
    be matched in order to really perform the action. It is evaluated
    on the current message that matched the rule.
  * rate: <num>/<period> specifies how much messages are to be
    generated (num), in the specified time period (period). Excess messages
    are dropped. For example: "1/60" allows 1 message / minute. Rates
    apply to the given scope for the given rule/action. E.g.
    context-scope="host", rate="1/60" means that one message gets generated
    for _each host_ per minute.
  * trigger: specifies when to execute the action, there are two possible
    triggers right now:
      * match: execute immediately once the rule matches
      * timeout: execute when the correllation timer expires

Messages that are generated by patterndb are emitted by the
internal() source for now. THIS IS TO BE CHANGED IN THE FUTURE.