radix: fixed incorrectly populated dbg_info
pdbtool match --debug-pattern uses the debug information to show which
portions of the pattern matches what in the incoming string. In case a
lookup fails because the parser doesn't match an incorrect debug entry
remained on the list, which is now dropped.

This may have caused string reads past the EOS character when displaying
the --debug-pattern output.

Reported-By: Peter Czanik <czanik@balabit.hu>