dbparser: watch the system time to move the correllation engine's idea of the current time
The correllation engine has an independent idea of the current time than
the system time, as it watches incoming messages and their timestamps
to move its own idea of the current time.

However if there are no incoming messages, time stood still until now.

This patch basically adds watching the current system time to update the
correllation's idea of the time.