patterndb: use unsigned char internally to avoid signed/unsigned comparisons
This patch actually fixes a concrete utf8 handling problem, but goes further
than that and changes all data structures to use unsigned chars.

This caused a large amount of warnings, which I've silenced using
-Wno-pointer-sign, which I consider suboptimal, but adding a lot of
scattered casts was even less nice.

The original patch was created by the PE team, which
I reworked when integrating to the OSE tree.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>