LogPipe: implement soft flow control mode
Soft flow control is used when the destination includes a local file and
the user doesn't expect to lose messages. It is basically equivalent to
hard flow control, but in case the destination becomes unwritable
(e.g. disk full) messages can get lost.

Currently it is not a user-visible feature (e.g. this cannot be enabled
by the administrator), but file destinations automatically use this.