config: close connections belonging to a stale source after SIGHUP
When syslog-ng is reloaded, the connections of a given source are
kept around (unless keep-alive() is no). However, some sources
may be removed from the configuration and the connections belonging
to these source became stale. They should have been closed, but in reality
they were not, because this code got broke somewhere in 3.0.

In case a message comes in on such a connection, syslog-ng crashes.

Reported-By: Zhengxiang Pan
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>