log statements: allow several semicolons instead of one
In order to introduce block support within "log" statements, it is convinient
to allow several ';' everywhere where only one was permitted.

This is needed in order to make it possible to close a statement inside
a block with a ';', and then also use a ';' at the end of a block reference.

This is more readable.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>