configure: fixed the export of the libexecdir directory if overridden with a configure option
Earlier, libexecdir was constructed in syslog-ng.h using the value of
the PATH_PREFIX macro, exported by the configure script.

This works OK as long as exec_prefix is not overridden on the
configure command line.

This patch only changes the instalation layout if exec_prefix is actually
specified and it has only an effect if --enable-env-wrapper is also used.

Reported-By: Matthew Hall
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>