dbparser: don't initialize the same db-parser() instance multiple times
Even if the configuration references the same db-parser() from several
locations.

This caused a difficult to understand config initialization problem in
syslog-ng.

NOTE: the current behaviour of the db-parsers() correllation state is going
to change in 3.3. In 3.2, the state is shared in this case, in 3.3 each
reference will use a private, independent state.

Reported-By: Matthew Hall
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>