LogSource: don't wait for window to empty in case the source is not using worker threads
In order to keep the messages in "flow", LogSource may add a little
nanosleep() in case the window is full. This is performed in the hope that
a minor sleep is enough to get some elements out of the window and a
sleep is much cheaper than going back to the main loop.

However this sleep() was introduced unconditionally for both threaded and
non-threaded sources, which is certainly not the best. Now the threaded
value is passed as an argument to log_source_set_options(). The sleeps
won't be performed if we're in a non-threaded source.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>