affile: fixed bogus permission bits on logfiles on platforms where mode_t is 16 bits only
syslog-ng is creating files with bogus permissions on platforms where mode_t
is 16 bits only. Such a beast is kFreeBSD.

Credits to the Debian folks for noticing & tracking down the issue.
The original reporter is in the "Reported-By" line, but also thanks for
gcs@debian.org, algernon@debian.org and peter.czanik@balabit.hu for
taking care of the issue.

Reported-By: Steven Chamberlain <steven@pyro.eu.org>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>