patterndb: keep correllation state when the database is reloaded
Earlier the correllation state was forgotten when the patterndb database
was reloaded, which happens automatically if the file is changed. This
is certainly a grave problem and this patch fixes that.

Since this was an API limitation, this patch refactores some of the
patterndb data structures in order to avoid having to reallocate PatternDB
on reload. Instead that detail is handled inside PatternDB.

Also, this moves some of the internal API to a separate header file.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>