afmongodb: Client library import.
With the changes below, we embed the libmongo-client library into
syslog-ng gracefully, using git submodules.

This means that after a fresh clone, one will need to run git
submodule update --init --recursive, which will initialise and update
every registered submodule, and check them out.

The build system was updated to both run configure in the new
submodule located at modules/afmongodb/libmongo-client, and to enter
said directory during a build.

Once the submodules are checked out, make all, install, check, dist
and the rest work as if these modules were integral parts of the
distribution.

In the future, if more submodules are added, one needs to add them to
autogen.sh (if appropriate) and to configure.in's AC_CONFIG_SUBDIRS().

Signed-off-by: Gergely Nagy <algernon@balabit.hu>