afmongodb: Made the included client library optional.
Added an --enable-included-mongo-client option to configure. When
specified (the default), the build will try to configure and build the
included library. When turned off, the system library will be used.

However, the system library is not checked for yet, because it doesn't
have a pkg-config file yet.

Adding such a check, however, is as easy as adding the appropriate
line in the else branch just before the AM_CONDITIONAL that sets
ENABLE_INCLUDED_MONGO_CLIENT.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>