afmongodb: lock the queue_mutex every time we change the queue.
Though log_queue should be locking itself, and the pop / push
functions indeed do, there is no such guarantee for
log_queue_get_length().

This means that unless we hold the same lock for all three operations,
there will be a race condition, and this is why deadlocks could
happen.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>