fixed segfault filter() expression evaluation
This patch fixes a segmentation fault and improves the performance
of filter() evaluation. Earlier the filter() expression always did a hash
table lookup in the fast path, instead of doing the lookup once and
reusing the result.

The patch introduces an "init" virtual function for filter expression nodes
that gets called when the configuration is initialized, e.g. when all
the filter expressions become available. This callback is then used
to do the lookup.

This way, the referenced filter rule need not be defined by the time
we reference it, it can come later in the configuration just as earlier.

Reported-By: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>